requirements.txt
ansi==0.3.6
APScheduler==3.9.1
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
cryptography==38.0.1
emoji==2.1.0
flake8==5.0.4
Flask-Mail==0.9.1
Flask==2.1.2
Flask-Login==0.6.2
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==1.1.3
idna==3.4
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mccabe==0.7.0
numpy==1.23.3
ntplib==0.4.0
pandas==1.5.0
pycodestyle==2.9.1
pyflakes==2.5.0
python-dateutil==2.8.2
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
six==1.16.0
SQLAlchemy==1.4.41
tzdata==2022.4
tzlocal==4.2
urllib3==1.26.12
Werkzeug==2.2.2
WTForms==3.0.1
yapf==0.32.0
zipp==3.8.1
